badge_base_url is a string or callable that can process RequestHandler
Instead of using a new setting launch_host_allowed reuse the badge_base_url setting by allowing it to be a callable that is passed the current RequestHandler. It can then be configured to extract the badge_base_url in any way, for example by looking for a X-Binder-Launch-Host header or a query parameter.